+++ src/pcm_resample.c
@@ -78,7 +78,7 @@ void pcm_resample_deinit(struct pcm_resample_state *state)
 
 const int16_t *
 pcm_resample_16(struct pcm_resample_state *state,
-		uint8_t channels,
+		unsigned channels,
 		unsigned src_rate, const int16_t *src_buffer, size_t src_size,
 		unsigned dest_rate, size_t *dest_size_r,
 		GError **error_r)
@@ -100,7 +100,7 @@ pcm_resample_16(struct pcm_resample_state *state,
 
 const int32_t *
 pcm_resample_32(struct pcm_resample_state *state,
-		uint8_t channels,
+		unsigned channels,
 		unsigned src_rate, const int32_t *src_buffer, size_t src_size,
 		unsigned dest_rate, size_t *dest_size_r,
 		GError **error_r)
