+++ src/pcm_resample.h
@@ -45,7 +45,7 @@ struct pcm_resample_state {
 	struct {
 		unsigned src_rate;
 		unsigned dest_rate;
-		uint8_t channels;
+		unsigned channels;
 	} prev;
 
 	int error;
@@ -82,7 +82,7 @@ void pcm_resample_deinit(struct pcm_resample_state *state);
  */
 const int16_t *
 pcm_resample_16(struct pcm_resample_state *state,
-		uint8_t channels,
+		unsigned channels,
 		unsigned src_rate,
 		const int16_t *src_buffer, size_t src_size,
 		unsigned dest_rate, size_t *dest_size_r,
@@ -102,7 +102,7 @@ pcm_resample_16(struct pcm_resample_state *state,
  */
 const int32_t *
 pcm_resample_32(struct pcm_resample_state *state,
-		uint8_t channels,
+		unsigned channels,
 		unsigned src_rate,
 		const int32_t *src_buffer, size_t src_size,
 		unsigned dest_rate, size_t *dest_size_r,
@@ -122,7 +122,7 @@ pcm_resample_32(struct pcm_resample_state *state,
  */
 static inline const int32_t *
 pcm_resample_24(struct pcm_resample_state *state,
-		uint8_t channels,
+		unsigned channels,
 		unsigned src_rate,
 		const int32_t *src_buffer, size_t src_size,
 		unsigned dest_rate, size_t *dest_size_r,
