+++ src/pcm_resample_libsamplerate.c
@@ -106,7 +106,7 @@ pcm_resample_lsr_deinit(struct pcm_resample_state *state)
 
 static bool
 pcm_resample_set(struct pcm_resample_state *state,
-		 uint8_t channels, unsigned src_rate, unsigned dest_rate,
+		 unsigned channels, unsigned src_rate, unsigned dest_rate,
 		 GError **error_r)
 {
 	int error;
@@ -159,7 +159,7 @@ lsr_process(struct pcm_resample_state *state, GError **error_r)
 
 const int16_t *
 pcm_resample_lsr_16(struct pcm_resample_state *state,
-		    uint8_t channels,
+		    unsigned channels,
 		    unsigned src_rate,
 		    const int16_t *src_buffer, size_t src_size,
 		    unsigned dest_rate, size_t *dest_size_r,
@@ -223,7 +223,7 @@ src_float_to_int_array (const float *in, int *out, int len)
 
 const int32_t *
 pcm_resample_lsr_32(struct pcm_resample_state *state,
-		    uint8_t channels,
+		    unsigned channels,
 		    unsigned src_rate,
 		    const int32_t *src_buffer, size_t src_size,
 		    unsigned dest_rate, size_t *dest_size_r,
